1. Added bsg_cuda_lite_runtime.h into software/bsg_manycore_lib 2. Changed the main.c in software/spmd/bsg_cuda_lite_runtime/main/ to use this library
